More 64-bit uint vs size_t fixes

+++ tango/io/selector/PollSelector.d
@@ -377,7 +377,7 @@ version (Posix)
             this.keys = keys;
         }
 
-        uint length()
+        size_t length()
         {
             return numSelected;
         }
